Moved PageQueryBuilder to page.ts
